Final implementation of the plot
