--- results/flatsat/requirements.txt
@@ -1,23 +1,0 @@
-contourpy==1.1.1
-cycler==0.12.0
-fonttools==4.43.0
-imageio==2.31.4
-kiwisolver==1.4.5
-lazy_loader==0.3
-matplotlib==3.8.0
-networkx==3.1
-numpy==1.26.0
-opencv-python==4.8.1.78
-packaging==23.1
-palettable==3.3.3
-pandas==2.1.1
-Pillow==10.0.1
-pyparsing==3.1.1
-python-dateutil==2.8.2
-pytz==2023.3.post1
-PyWavelets==1.4.1
-scikit-image==0.21.0
-scipy==1.11.3
-six==1.16.0
-tifffile==2023.9.26
-tzdata==2023.3
